refactor(store,wire,core,expr,sql,server): the ORDER BY comparator, the OPK byte container and the index-key promotion each have one spelling
`cmp_order_keys` moves to `gnitz-expr` beside `ColumnLocator`: the worker's
top-k and the client's ordering sink now read one comparator instead of two
held equivalent by a doc comment, and the replica that pinned them against
each other goes with it.

`PkBuf` moves to `gnitz-wire`, beside `MAX_PK_BYTES` and the per-column codec
whose output it carries; `PkTuple` is deleted and its schema-aware
constructors become free functions beside the client's own native<->OPK walk,
which they now share rather than duplicate. The buffer's writers are
`append` / `edit` / `try_write`, each deriving the width tag from the span it
hands out, so `set_len` is private and no writer can report a width it did not
write. `Borrow<[u8]>` drops a per-row key mint from the transactional
UPDATE/DELETE overlay.

`IndexKeySpec` owns the index-key promotion and yields the schema its entries
land in, so `make_index_schema` is that call; the two sites that built both
from identical inputs build one.

Batch construction stops copying the layout arrays by value: `compute_offsets`
takes an out-parameter, the five VM-path constructors take `&SchemaDescriptor`,
and `SysFamily::schema()` borrows its static. The VM dispatch loop loses 460
instructions and four of each large `memcpy` immediate.

`locate` is one walk of the PK list, and `pk_byte_offset` — its only caller's
only caller, carrying the module's only `unreachable!()` — is gone.

Guards: the relay route's client-supplied columns are bounded in
`prepare_relay` through the compiler's own `oob_cols`; an out-of-range
distribution prefix is rejected at both decode boundaries instead of clamped;
a float reindex or group key is refused rather than packed into a slot where
+0.0 and -0.0 differ byte-wise while comparing equal; and `SchemaBlock::decode`
sums the PK stride it was leaving to a coincidence between MAX_PK_BYTES and
MAX_PK_COLUMNS.

+++ crates/gnitz-server/src/query/vm/tests/exec.rs
@@ -175,7 +175,7 @@ fn a_union_with_an_empty_left_operand_returns_the_right_one() {
 
     assert_eq!(extract_rows(&result), vec![(3, 2, 30), (4, -1, 40)]);
     assert_eq!(
-        result.pk_stride(),
+        usize::from(result.pk_stride()),
         schema.pk_stride(),
         "the output carries the register's own shape",
     );
